pyproject.toml
[project]
name = "psy-agent"
version = "0.1.0"
description = "AI Psychotherapy Agent"
authors = [
    { name = "Vitaliy Sharandin", email = "vitaliy.sharandin@gmail.com" },
]
requires-python = ">=3.12"

dependencies = [
    "streamlit==1.42.1",
    "chromadb==0.5.23",
    "llama-index==0.12.7",
    "langchain==0.3.22",
    "langchain-openai==0.3.12",
    "langchain-ollama==0.3.0",
    "langgraph-checkpoint-sqlite==2.0.1",
    "llama-index-embeddings-ollama==0.5.0",
    "llama-index-vector-stores-chroma==0.4.1",
    "llama-index-llms-ollama==0.5.0",
    "llama-index-postprocessor-flag-embedding-reranker==0.3.0",
    "langgraph==0.3.27",
    "tavily-python==0.5.0",
    "prometheus-client==0.21.1",
    "langfuse==2.59.3",
    "pytest==8.3.5",
    "ragas==0.2.14",
    "llama-index-llms-langchain==0.6.1",
    "deepeval==2.7.5",
    "pytest-mock==3.14.0",
    "mcp==1.9.3",
    "langchain-mcp-adapters==0.1.7",
    "aiosqlite"
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["psy-agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff.format]
indent-style = "space"

[tool.ruff]
line-length = 120
